Added Components folder, moved NavItem into components folder
